[placemat] chdir to the directory of YAML file
In YAML, file paths are often written as relative from the YAML
itself.  If placemat is invoked from the directory other than the
directory of YAML file, such files cannot be located.

To allow such relative paths in YAML, this commit let placemat
chdir to the directory of the first YAML given in command-line.