Fix the permission error
QEMU raises the permission error "Cannot get 'write' permission without 'resize': Image size is not a multiple of request alignment",
if a target image is not 4k aligned, and WRITE permission is granted but RESIZE is not.
To work around the issue, we change the localds image fomat from raw to qcow2 not to generate an unaligned seed image.

See
https://github.com/qemu/qemu/commit/9c60a5d1978e6dcf85c0e01b50e6f7f54ca09104
https://bugs.launchpad.net/ubuntu/+source/qemu/+bug/1921665

Signed-off-by: Yusuke Suzuki <yusuke-suzuki@cybozu.co.jp>